+++ libsolidity/codegen/CompilerUtils.cpp
@@ -98,16 +98,17 @@ void CompilerUtils::revertWithStringData(Type const& _argumentType)
 void CompilerUtils::accessCalldataTail(Type const& _type)
 {
 	solAssert(_type.dataStoredIn(DataLocation::CallData), "");
+	solAssert(_type.isDynamicallyEncoded(), "");
 
-	unsigned int baseEncodedSize = _type.calldataEncodedSize();
-	solAssert(baseEncodedSize > 1, "");
+	unsigned int tailSize = _type.calldataEncodedTailSize();
+	solAssert(tailSize > 1, "");
 
 	// returns the absolute offset of the tail in "base_ref"
 	m_context.appendInlineAssembly(Whiskers(R"({
 		let rel_offset_of_tail := calldataload(ptr_to_tail)
 		if iszero(slt(rel_offset_of_tail, sub(sub(calldatasize(), base_ref), sub(<neededLength>, 1)))) { revert(0, 0) }
 		base_ref := add(base_ref, rel_offset_of_tail)
-	})")("neededLength", toCompactHexWithPrefix(baseEncodedSize)).render(), {"base_ref", "ptr_to_tail"});
+	})")("neededLength", toCompactHexWithPrefix(tailSize)).render(), {"base_ref", "ptr_to_tail"});
 	// stack layout: <absolute_offset_of_tail> <garbage>
 
 	if (!_type.isDynamicallySized())
@@ -170,7 +171,7 @@ void CompilerUtils::loadFromMemoryDynamic(
 		solAssert(!_fromCalldata, "");
 		solAssert(_padToWordBoundaries, "");
 		if (_keepUpdatedMemoryOffset)
-			m_context << arrayType->memorySize() << Instruction::ADD;
+			m_context << arrayType->memoryDataSize() << Instruction::ADD;
 	}
 	else
 	{
@@ -251,7 +252,7 @@ void CompilerUtils::abiDecode(TypePointers const& _typeParameters, bool _fromMem
 	//@todo this does not yet support nested dynamic arrays
 	size_t encodedSize = 0;
 	for (auto const& t: _typeParameters)
-		encodedSize += t->decodingType()->calldataEncodedSize(true);
+		encodedSize += t->decodingType()->calldataHeadSize();
 	m_context.appendInlineAssembly("{ if lt(len, " + to_string(encodedSize) + ") { revert(0, 0) } }", {"len"});
 
 	m_context << Instruction::DUP2 << Instruction::ADD;
@@ -321,7 +322,7 @@ void CompilerUtils::abiDecode(TypePointers const& _typeParameters, bool _fromMem
 					// Size has already been checked for this one.
 					moveIntoStack(2);
 					m_context << Instruction::DUP3;
-					m_context << u256(arrayType.calldataEncodedSize(true)) << Instruction::ADD;
+					m_context << u256(arrayType.calldataHeadSize()) << Instruction::ADD;
 				}
 			}
 			else
@@ -358,7 +359,7 @@ void CompilerUtils::abiDecode(TypePointers const& _typeParameters, bool _fromMem
 					// size has already been checked
 					// stack: input_end base_offset data_offset
 					m_context << Instruction::DUP1;
-					m_context << u256(calldataType->calldataEncodedSize()) << Instruction::ADD;
+					m_context << u256(calldataType->calldataHeadSize()) << Instruction::ADD;
 				}
 				if (arrayType.location() == DataLocation::Memory)
 				{
@@ -588,11 +589,6 @@ void CompilerUtils::zeroInitialiseMemoryArray(ArrayType const& _type)
 	}
 	else
 	{
-		// TODO: Potential optimization:
-		// When we create a new multi-dimensional dynamic array, each element
-		// is initialized to an empty array. It actually does not hurt
-		// to re-use exactly the same empty array for all elements. Currently,
-		// a new one is created each time.
 		auto repeat = m_context.newTag();
 		m_context << repeat;
 		pushZeroValue(*_type.baseType());
@@ -1014,7 +1010,7 @@ void CompilerUtils::convertType(
 				{
 					CompilerUtils utils(_context);
 					// stack: <source ref>
-					utils.allocateMemory(typeOnStack->memorySize());
+					utils.allocateMemory(typeOnStack->memoryDataSize());
 					_context << Instruction::SWAP1 << Instruction::DUP2;
 					// stack: <memory ptr> <source ref> <memory ptr>
 					for (auto const& member: typeOnStack->members(nullptr))
@@ -1197,7 +1193,8 @@ void CompilerUtils::pushZeroValue(Type const& _type)
 		1,
 		[type](CompilerContext& _context) {
 			CompilerUtils utils(_context);
-			utils.allocateMemory(max(32u, type->calldataEncodedSize()));
+
+			utils.allocateMemory(max<u256>(32u, type->memoryDataSize()));
 			_context << Instruction::DUP1;
 
 			if (auto structType = dynamic_cast<StructType const*>(type))
@@ -1364,6 +1361,8 @@ void CompilerUtils::storeStringData(bytesConstRef _data)
 
 unsigned CompilerUtils::loadFromMemoryHelper(Type const& _type, bool _fromCalldata, bool _padToWords)
 {
+	solAssert(_type.isValueType(), "");
+
 	unsigned numBytes = _type.calldataEncodedSize(_padToWords);
 	bool isExternalFunctionType = false;
 	if (auto const* funType = dynamic_cast<FunctionType const*>(&_type))
@@ -1435,6 +1434,8 @@ unsigned CompilerUtils::prepareMemoryStore(Type const& _type, bool _padToWords)
 		_type.sizeOnStack() == 1,
 		"Memory store of types with stack size != 1 not allowed (Type: " + _type.toString(true) + ")."
 	);
+
+	solAssert(!_type.isDynamicallyEncoded(), "");
 
 	unsigned numBytes = _type.calldataEncodedSize(_padToWords);