+++ libsolidity/codegen/YulUtilFunctions.cpp
@@ -652,11 +652,11 @@ string YulUtilFunctions::arrayConvertLengthToSize(ArrayType const& _type)
 						<?byteArray>
 							size := length
 						<!byteArray>
-							size := <mul>(length, <elementSize>)
+							size := <mul>(length, <stride>)
 						</byteArray>
 					})")
 					("functionName", functionName)
-					("elementSize", to_string(_type.location() == DataLocation::Memory ? baseType.memoryHeadSize() : baseType.calldataEncodedSize()))
+					("stride", to_string(_type.location() == DataLocation::Memory ? _type.memoryStride() : _type.calldataStride()))
 					("byteArray", _type.isByteArray())
 					("mul", overflowCheckedIntMulFunction(*TypeProvider::uint256()))
 					.render();
@@ -812,10 +812,7 @@ string YulUtilFunctions::nextArrayElementFunction(ArrayType const& _type)
 		}
 		case DataLocation::CallData:
 		{
-			u256 size =
-				_type.baseType()->isDynamicallyEncoded() ?
-				32 :
-				_type.baseType()->calldataEncodedSize();
+			u256 size = _type.calldataStride();
 			solAssert(size >= 32 && size % 32 == 0, "");
 			templ("advance", toCompactHexWithPrefix(size));
 			break;